new app& Image composer divide
