--- install_log.txt
@@ -1,4 +1,0 @@
-----------------------------
-Date: 2025-06-25 04:50:10
-Error Message: Failed to install Scoop and Fastfetch.
-----------------------------
